+++ package/tests/test_qt_part/test_settings.py
@@ -1,11 +1,8 @@
-import itertools
 import os
 import typing
-from copy import deepcopy
 
 import numpy as np
 import pytest
-import SimpleITK as sitk
 
 
 from PartSeg.common_backend.base_settings import BaseSettings
@@ -16,126 +13,10 @@ from PartSegCore.analysis import analysis_algorithm_dict
 from PartSegCore.analysis.io_utils import create_history_element_from_project
 from PartSegCore.analysis.load_functions import LoadProject
 from PartSegCore.analysis.save_functions import SaveProject
-from PartSegCore.image_operations import RadiusType
-from PartSegCore.io_utils import HistoryElement, HistoryProblem
+from PartSegCore.io_utils import HistoryProblem
 from PartSegCore.mask.history_utils import create_history_element_from_segmentation_tuple
-from PartSegCore.mask.io_functions import LoadSegmentationImage, LoadStackImage, SegmentationTuple
-from PartSegCore.mask_create import calculate_mask, MaskProperty
-from PartSegImage import Image
-
-
-@pytest.fixture
-def image():
-    data = np.zeros([20, 20, 20, 2], dtype=np.uint8)
-    data[10:-1, 1:-1, 1:-1, 0] = 20
-    data[1:10, 1:-1, 1:-1, 1] = 20
-    data[1:-1, 1:5, 1:-1, 1] = 20
-    data[1:-1, -5:-1, 1:-1, 1] = 20
-
-    return Image(data, (1, 1, 1), axes_order="ZYXC")
-
-
-@pytest.fixture
-def stack_image():
-    data = np.zeros([20, 40, 40], dtype=np.uint8)
-    for x, y in itertools.product([0, 20], repeat=2):
-        data[1:-1, x + 2 : x + 18, y + 2 : y + 18] = 100
-    for x, y in itertools.product([0, 20], repeat=2):
-        data[3:-3, x + 4 : x + 16, y + 4 : y + 16] = 120
-    for x, y in itertools.product([0, 20], repeat=2):
-        data[5:-5, x + 6 : x + 14, y + 6 : y + 14] = 140
-
-    return SegmentationTuple("test_path", Image(data, (2, 1, 1), axes_order="ZYX", file_path="test_path"))
-
-
-@pytest.fixture
-def algorithm_parameters():
-    algorithm_parameters = {
-        "algorithm_name": "Lower threshold",
-        "values": {
-            "threshold": {"name": "Manual", "values": {"threshold": 10}},
-            "channel": 0,
-            "noise_filtering": {"name": "None", "values": {}},
-            "minimum_size": 1,
-            "side_connection": False,
-        },
-    }
-    return deepcopy(algorithm_parameters)
-
-
-@pytest.fixture
-def mask_segmentation_parameters():
-    return {
-        "algorithm": "Threshold",
-        "values": {
-            "channel": 0,
-            "noise_filtering": {"name": "None", "values": {}},
-            "threshold": {"name": "Manual", "values": {"threshold": 10}},
-            "close_holes": False,
-            "close_holes_size": 200,
-            "smooth_border": {"name": "None", "values": {}},
-            "side_connection": False,
-            "minimum_size": 10000,
-            "use_convex": False,
-        },
-    }
-
-
-@pytest.fixture
-def stack_segmentation1(stack_image: SegmentationTuple, mask_segmentation_parameters):
-    data = np.zeros([20, 40, 40], dtype=np.uint8)
-    for i, (x, y) in enumerate(itertools.product([0, 20], repeat=2), start=1):
-        data[1:-1, x + 2 : x + 18, y + 2 : y + 18] = i
-    parameters = {i: deepcopy(mask_segmentation_parameters) for i in range(1, 5)}
-    return stack_image._replace(segmentation=data, segmentation_parameters=parameters, selected_components=[1, 3])
-
-
-@pytest.fixture
-def stack_segmentation2(stack_image: SegmentationTuple, mask_segmentation_parameters):
-    data = np.zeros([20, 40, 40], dtype=np.uint8)
-    for i, (x, y) in enumerate(itertools.product([0, 20], repeat=2), start=1):
-        data[3:-3, x + 4 : x + 16, y + 4 : y + 16] = i
-    mask_segmentation_parameters["values"]["threshold"]["values"]["threshold"] = 110
-    parameters = {i: deepcopy(mask_segmentation_parameters) for i in range(1, 5)}
-    return stack_image._replace(segmentation=data, segmentation_parameters=parameters, selected_components=[1, 3])
-
-
-@pytest.fixture
-def mask_property():
-    return MaskProperty(RadiusType.NO, 0, RadiusType.NO, 0, False, False, False)
-
-
-@pytest.fixture(scope="module")
-def nucleus_project(data_test_dir):
-    # TODO change on synthetic example
-    return LoadStackImage.load([os.path.join(data_test_dir, "test_nucleus.tif")])
-
-
-@pytest.fixture
-def segmentation_fun(nucleus_project, mask_segmentation_parameters):
-    channel = nucleus_project.image.get_channel(0)[0]
-    threshold_img = (channel > 8000).astype(np.uint8)
-    components = sitk.GetArrayFromImage(
-        sitk.RelabelComponent(sitk.ConnectedComponent(sitk.GetImageFromArray(threshold_img)), 200)
-    )
-
-    def calculate(threshold=None):
-        if threshold is not None:
-            if threshold < 8000:
-                raise ValueError("Not supported")
-
-            threshold_img = channel > threshold
-            comp = np.copy(components)
-            comp[threshold_img == 0] = 0
-        else:
-            comp = components
-
-        return nucleus_project.replace_(
-            segmentation=comp, segmentation_parameters={i: deepcopy(mask_segmentation_parameters) for i in range(1, 6)}
-        )
-
-    return calculate
-
+from PartSegCore.mask.io_functions import LoadStackImage
+from PartSegCore.mask_create import calculate_mask
 
 @pytest.fixture
 def stack_settings(qtbot, tmp_path):
